refactor(ui/menu): replace menuV2 with menu
