+++ packages/varlet-cli/lib/index.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 "use strict";
 var __importDefault = (this && this.__importDefault) || function (mod) {
-  return (mod && mod.__esModule) ? mod : { "default": mod };
+    return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const logger_1 = __importDefault(require("./shared/logger"));
@@ -21,64 +21,64 @@ const commitLint_1 = require("./commands/commitLint");
 const program = new commander_1.Command();
 program.version(`varlet-cli ${require('../package.json').version}`).usage('<command> [options]');
 program
-  .command('dev')
-  .option('-f --force', 'Force dep pre-optimization regardless of whether deps have changed')
-  .description('Run varlet development environment')
-  .action(dev_1.dev);
+    .command('dev')
+    .option('-f --force', 'Force dep pre-optimization regardless of whether deps have changed')
+    .description('Run varlet development environment')
+    .action(dev_1.dev);
 program.command('build').description('Build varlet site for production').action(build_1.build);
 program
-  .command('build:vite')
-  .description('Use vite build app for production')
-  .action(() => (0, vite_1.vite)('build'));
+    .command('build:vite')
+    .description('Use vite build app for production')
+    .action(() => (0, vite_1.vite)('build'));
 program
-  .command('dev:vite')
-  .description('Use vite start server for development')
-  .action(() => (0, vite_1.vite)('dev'));
+    .command('dev:vite')
+    .description('Use vite start server for development')
+    .action(() => (0, vite_1.vite)('dev'));
 program.command('preview').description('Preview varlet site for production').action(preview_1.preview);
 program
-  .command('compile')
-  .description('Compile varlet components library code')
-  .option('-nu, --noUmd', 'Do not compile umd target code')
-  .action(compile_1.compile);
+    .command('compile')
+    .description('Compile varlet components library code')
+    .option('-nu, --noUmd', 'Do not compile umd target code')
+    .action(compile_1.compile);
 program.command('lint').description('Lint code').action(lint_1.lint);
 program
-  .command('create')
-  .description('Create a component directory')
-  .option('-n, --name <componentName>', 'Component name')
-  .option('-s, --sfc', 'Generate files in sfc format')
-  .option('-t, --tsx', 'Generate files in tsx format')
-  .option('-l, --locale', 'Generator internationalized files')
-  .action(create_1.create);
+    .command('create')
+    .description('Create a component directory')
+    .option('-n, --name <componentName>', 'Component name')
+    .option('-s, --sfc', 'Generate files in sfc format')
+    .option('-t, --tsx', 'Generate files in tsx format')
+    .option('-l, --locale', 'Generator internationalized files')
+    .action(create_1.create);
 program
-  .command('test')
-  .description('Run test in work directory')
-  .option('-w, --watch', 'Watch files for changes and rerun tests related to changed files')
-  .option('-c, --component <componentName>', 'Test a specific component')
-  .option('-cov, --coverage', 'Generate the coverage')
-  .action(test_1.test);
+    .command('test')
+    .description('Run test in work directory')
+    .option('-w, --watch', 'Watch files for changes and rerun tests related to changed files')
+    .option('-c, --component <componentName>', 'Test a specific component')
+    .option('-cov, --coverage', 'Generate the coverage')
+    .action(test_1.test);
 program
-  .command('gen')
-  .description('Generate cli application')
-  .option('-n, --name <applicationName>', 'Application name')
-  .option('-s, --sfc', 'Generate files in sfc format')
-  .option('-t, --tsx', 'Generate files in tsx format')
-  .option('-l, --locale', 'Generator internationalized files')
-  .action(gen_1.gen);
+    .command('gen')
+    .description('Generate cli application')
+    .option('-n, --name <applicationName>', 'Application name')
+    .option('-s, --sfc', 'Generate files in sfc format')
+    .option('-t, --tsx', 'Generate files in tsx format')
+    .option('-l, --locale', 'Generator internationalized files')
+    .action(gen_1.gen);
 program
-  .command('changelog')
-  .option('-rc --releaseCount <releaseCount>', 'Release count')
-  .option('-f --file <file>', 'Changelog filename')
-  .description('Generate changelog')
-  .action(changelog_1.changelog);
+    .command('changelog')
+    .option('-rc --releaseCount <releaseCount>', 'Release count')
+    .option('-f --file <file>', 'Changelog filename')
+    .description('Generate changelog')
+    .action(changelog_1.changelog);
 program
-  .command('release')
-  .option('-r --remote <remote>', 'Remote name')
-  .description('Release all packages and generate changelogs')
-  .action(release_1.release);
+    .command('release')
+    .option('-r --remote <remote>', 'Remote name')
+    .description('Release all packages and generate changelogs')
+    .action(release_1.release);
 program.command('commit-lint <gitParams>').description('Lint commit message').action(commitLint_1.commitLint);
 program.on('command:*', ([cmd]) => {
-  program.outputHelp();
-  logger_1.default.error(`\nUnknown command ${cmd}.\n`);
-  process.exitCode = 1;
+    program.outputHelp();
+    logger_1.default.error(`\nUnknown command ${cmd}.\n`);
+    process.exitCode = 1;
 });
 program.parse();
