+++ packages/varlet-ui/src/menu/example/index.vue
@@ -2,154 +2,20 @@
 import VarMenu from '..'
 import VarButton from '../../button'
 import VarCell from '../../cell'
-import Snackbar from '../../snackbar'
-import AppType from '@varlet/cli/site/mobile/components/AppType'
-import { reactive, toRefs, ref } from 'vue'
-import { pack, use } from './locale'
-import { watchLang, watchDarkMode } from '@varlet/cli/site/utils'
-import dark from '../../themes/dark'
+import dark from '../../themes/dark/index'
+import { watchDarkMode } from '@varlet/cli/site/utils'
 
-const values = reactive({
-  top: false,
-  bottom: false,
-  offsetX: false,
-  offsetX1: false,
-  offsetY: false,
-  offsetY1: false,
-  event: false,
-})
-const bgColor = ref('#fff')
-const { top, bottom, offsetX, offsetX1, offsetY, offsetY1, event } = toRefs(values)
-
-watchLang(use)
-watchDarkMode(dark, (theme) => {
-  bgColor.value = theme === 'darkTheme' ? '#272727' : '#fff'
-})
+watchDarkMode(dark)
 </script>
 
 <template>
-  <app-type>{{ pack.alignmentMethods }}</app-type>
-  <div class="block-1">
-    <var-menu v-model:show="top">
-      <var-button type="primary" @click="top = true">{{ pack.topAlignment }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-  </div>
-
-  <div class="block">
-    <var-menu alignment="bottom" v-model:show="bottom">
-      <var-button type="primary" @click="bottom = true">{{ pack.bottomAlignment }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-  </div>
-
-  <app-type>{{ pack.offset }}</app-type>
-
-  <div class="block-1">
-    <var-menu :offset-x="72" v-model:show="offsetX">
-      <var-button type="primary" @click="offsetX = true">{{ pack.offsetRight }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-
-    <var-menu :offset-x="-72" v-model:show="offsetX1">
-      <var-button type="primary" @click="offsetX1 = true">{{ pack.offsetLeft }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-  </div>
-
-  <div class="block-2">
-    <var-menu :offset-y="36" v-model:show="offsetY">
-      <var-button type="primary" @click="offsetY = true">{{ pack.offsetBottom }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-
-    <var-menu :offset-y="-36" v-model:show="offsetY1">
-      <var-button type="primary" @click="offsetY1 = true">{{ pack.offsetTop }}</var-button>
-
-      <template #menu>
-        <div class="cell-list" :style="{ background: bgColor }">
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-          <var-cell>{{ pack.menuOption }}</var-cell>
-        </div>
-      </template>
-    </var-menu>
-  </div>
-
-  <app-type>{{ pack.events }}</app-type>
-  <var-menu
-    v-model:show="event"
-    @open="() => Snackbar.info('open')"
-    @opened="() => Snackbar.success('opened')"
-    @close="() => Snackbar.warning('close')"
-    @closed="() => Snackbar.error('closed')"
-  >
-    <var-button type="primary" @click="event = true">{{ pack.events }}</var-button>
+  <var-menu>
+    <var-button type="primary">触发</var-button>
 
     <template #menu>
-      <div class="cell-list" :style="{ background: bgColor }">
-        <var-cell>{{ pack.menuOption }}</var-cell>
-        <var-cell>{{ pack.menuOption }}</var-cell>
-        <var-cell>{{ pack.menuOption }}</var-cell>
-      </div>
+      <var-cell>菜单项</var-cell>
+      <var-cell>菜单项</var-cell>
+      <var-cell>菜单项</var-cell>
     </template>
   </var-menu>
-
-  <div style="margin-bottom: 100px"></div>
 </template>
-
-<style scoped lang="less">
-.cell-list {
-  transition: background-color 0.25s;
-}
-
-.block {
-  margin-top: 130px;
-}
-
-.block-1 {
-  display: flex;
-  justify-content: space-between;
-}
-
-.block-2 {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 130px;
-}
-</style>
