--- packages/varlet-ui/src/menu/example/locale/en-US.ts
@@ -1,12 +1,0 @@
-export default {
-  alignmentMethods: 'Alignment Methods',
-  topAlignment: 'Top Alignment',
-  menuOption: 'Menu Option',
-  bottomAlignment: 'Bottom Alignment',
-  offset: 'Offset',
-  offsetRight: 'Offset Right',
-  offsetLeft: 'Offset Left',
-  offsetBottom: 'Offset Bottom',
-  offsetTop: 'Offset Top',
-  events: 'Events',
-}
