--- packages/varlet-ui/src/menu/example/locale/zh-CN.ts
@@ -1,12 +1,0 @@
-export default {
-  alignmentMethods: '对齐方式',
-  topAlignment: '顶部对齐',
-  menuOption: '菜单项',
-  bottomAlignment: '底部对齐',
-  offset: '偏移量',
-  offsetRight: '右偏移',
-  offsetLeft: '左偏移',
-  offsetBottom: '下偏移',
-  offsetTop: '上偏移',
-  events: '注册事件',
-}
