+++ packages/varlet-ui/src/menu/index.ts
@@ -1,5 +1,5 @@
 import type { App } from 'vue'
-import Menu from './Menu'
+import Menu from './Menu.vue'
 
 Menu.install = function (app: App) {
   app.component(Menu.name, Menu)
