+++ packages/varlet-ui/src/menu/menu.less
@@ -1,21 +1,25 @@
+:root {
+  --menu-background-color: #fff;
+}
+
 .var {
   &-menu-enter-from,
   &-menu-leave-to {
-    transform: scale(0.8);
     opacity: 0;
   }
+
   &-menu-enter-active,
   &-menu-leave-active {
-    transition-property: opacity, transform;
+    transition-property: opacity;
     transition-duration: 0.25s;
   }
 }
 
 .var-menu {
   display: inline-block;
-  transition: background-color 0.25s;
 
-  &__menu {
-    position: absolute;
+  &--menu-background-color {
+    background: var(--menu-background-color);
+    transition: opacity 0.25s, background-color 0.25s;
   }
 }
