+++ packages/varlet-ui/src/themes/dark/index.ts
@@ -28,7 +28,7 @@ import tabs from './tabs'
 import appBar from './appBar'
 import bottomNavigation from './bottomNavigation'
 import bottomNavigationItem from './bottomNavigationItem'
-import menuV2 from './menuV2'
+import menu from './menu'
 
 export default {
   // common
@@ -71,5 +71,5 @@ export default {
   ...appBar,
   ...bottomNavigation,
   ...bottomNavigationItem,
-  ...menuV2,
+  ...menu,
 } as StyleVars
