+++ packages/varlet-ui/src/themes/dark/menu.ts
@@ -1,0 +1,3 @@
+export default {
+  '--menu-background-color': '#272727',
+}
