--- packages/varlet-ui/src/themes/dark/menuV2.ts
@@ -1,3 +1,0 @@
-export default {
-  '--menu-v2-background-color': '#272727',
-}
